+++ src/deployCommand.ts
@@ -6,9 +6,14 @@
  */
 
 import * as fs from 'fs';
-import { ComponentSet, DeployResult, MetadataApiDeployStatus } from '@salesforce/source-deploy-retrieve';
+import {
+  ComponentSet,
+  DeployResult,
+  MetadataApiDeploy,
+  MetadataApiDeployStatus,
+} from '@salesforce/source-deploy-retrieve';
 import { ConfigAggregator, ConfigFile, PollingClient, SfdxError, StatusResult } from '@salesforce/core';
-import { AnyJson, asString, getBoolean } from '@salesforce/ts-types';
+import { AnyJson, getBoolean } from '@salesforce/ts-types';
 import { Duration, once } from '@salesforce/kit';
 import { SourceCommand } from './sourceCommand';
 
@@ -18,17 +23,19 @@ interface StashFile {
 }
 
 export abstract class DeployCommand extends SourceCommand {
-  protected static readonly STASH_KEY = 'SOURCE_DEPLOY';
+  protected static readonly SOURCE_STASH_KEY = 'SOURCE_DEPLOY';
+  protected static readonly MDAPI_STASH_KEY = 'MDAPI_DEPLOY';
 
   protected displayDeployId = once((id: string) => {
     if (!this.isJsonOutput()) {
       this.ux.log(`Deploy ID: ${id}`);
     }
   });
-
+  // used to determine the correct stash.json key
+  protected isSourceStash = true;
   protected deployResult: DeployResult;
   /**
-   * Request a report of an in-progess or completed deployment.
+   * Request a report of an in-progress or completed deployment.
    *
    * @param id the Deploy ID of a deployment request
    * @returns DeployResult
@@ -47,10 +54,25 @@ export abstract class DeployCommand extends SourceCommand {
   protected setStash(deployId: string): void {
     const file = this.getStash();
     this.logger.debug(`Stashing deploy ID: ${deployId} in ${file.getPath()}`);
-    file.writeSync({ [DeployCommand.STASH_KEY]: { jobid: deployId } });
+    file.writeSync({
+      [this.getStashKey()]: { jobid: deployId },
+    });
   }
 
-  protected resolveDeployId(id: string): string {
+  /**
+   * This method is here to provide a workaround to stubbing a constructor in the tests.
+   *
+   * @param id
+   */
+  protected createDeploy(id?: string): MetadataApiDeploy {
+    return new MetadataApiDeploy({ usernameOrConnection: this.org.getUsername(), id });
+  }
+
+  protected getStashKey(): string {
+    return this.isSourceStash ? DeployCommand.SOURCE_STASH_KEY : DeployCommand.MDAPI_STASH_KEY;
+  }
+
+  protected resolveDeployId(id?: string): string {
     let stash: ConfigFile<StashFile>;
     if (id) {
       return id;
@@ -58,13 +80,17 @@ export abstract class DeployCommand extends SourceCommand {
       try {
         stash = this.getStash();
         stash.readSync(true);
-        const deployId = asString((stash.get(DeployCommand.STASH_KEY) as { jobid: string }).jobid);
+        const deployId = (
+          stash.get(this.getStashKey()) as {
+            jobid: string;
+          }
+        ).jobid;
         this.logger.debug(`Using deploy ID: ${deployId} from ${stash.getPath()}`);
         return deployId;
       } catch (err: unknown) {
         const error = err as Error & { code: string };
         if (error.name === 'JsonParseError') {
-          const stashFilePath = stash.getPath();
+          const stashFilePath = stash?.getPath();
           const corruptFilePath = `${stashFilePath}_corrupted_${Date.now()}`;
           fs.renameSync(stashFilePath, corruptFilePath);
           const invalidStashErr = SfdxError.create('@salesforce/plugin-source', 'deploy', 'InvalidStashFile', [
@@ -74,7 +100,8 @@ export abstract class DeployCommand extends SourceCommand {
           invalidStashErr.stack = `${invalidStashErr.stack}\nDue to:\n${error.stack}`;
           throw invalidStashErr;
         }
-        if (error.code === 'ENOENT') {
+        if (error.code === 'ENOENT' || !stash?.get(this.getStashKey())) {
+          // if the file doesn't exist, or the key doesn't exist in the stash
           throw SfdxError.create('@salesforce/plugin-source', 'deploy', 'MissingDeployId');
         }
         throw SfdxError.wrap(error);