+++ test/nuts/mdapi.nut.ts
@@ -11,7 +11,9 @@ import { expect } from 'chai';
 import { TestSession, execCmd } from '@salesforce/cli-plugins-testkit';
 import { ComponentSet, SourceComponent } from '@salesforce/source-deploy-retrieve';
 import { DescribeMetadataResult } from 'jsforce';
+import { exec } from 'shelljs';
 import { ConvertCommandResult } from '../../src/formatters/mdapi/convertResultFormatter';
+import { DeployCancelCommandResult } from '../../src/formatters/deployCancelResultFormatter';
 
 let session: TestSession;
 
@@ -159,6 +161,38 @@ describe('mdapi NUTs', () => {
           expect(size1 / size2).to.be.within(0.98, 1.02);
         }
       }
+    });
+  });
+
+  describe('mdapi:deploy:cancel', () => {
+    it('will cancel an mdapi deploy via the stash.json', () => {
+      execCmd('force:source:convert --outputdir mdapi');
+      // TODO: once mdapi:deploy is migrated switch to execCmd
+      const deploy = JSON.parse(exec('sfdx force:mdapi:deploy -d mdapi -w 0 --json', { silent: true })) as {
+        result: { id: string };
+      };
+      const result = execCmd<DeployCancelCommandResult>('force:mdapi:deploy:cancel --json');
+      expect(result.jsonOutput.status).to.equal(0);
+      const json = result.jsonOutput.result;
+      expect(json).to.have.property('canceledBy');
+      expect(json).to.have.property('status');
+      expect(json.status).to.equal('Canceled');
+      expect(json.id).to.equal(deploy.result.id);
+    });
+
+    it('will cancel an mdapi deploy via the specified deploy id', () => {
+      execCmd('force:source:convert --outputdir mdapi');
+      // TODO: once mdapi:deploy is migrated switch to execCmd
+      const deploy = JSON.parse(exec('sfdx force:mdapi:deploy -d mdapi -w 0 --json', { silent: true })) as {
+        result: { id: string };
+      };
+      const result = execCmd<DeployCancelCommandResult>(`force:mdapi:deploy:cancel --json --jobid ${deploy.result.id}`);
+      expect(result.jsonOutput.status).to.equal(0);
+      const json = result.jsonOutput.result;
+      expect(json).to.have.property('canceledBy');
+      expect(json).to.have.property('status');
+      expect(json.status).to.equal('Canceled');
+      expect(json.id).to.equal(deploy.result.id);
     });
   });