+++ CHANGELOG.md
@@ -2,12 +2,23 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### [1.4.2](https://github.com/salesforcecli/plugin-source/compare/v1.4.1...v1.4.2) (2021-11-11)
+## [1.5.0](https://github.com/salesforcecli/plugin-source/compare/v1.4.3...v1.5.0) (2021-11-16)
 
+### Features
+
+- add mdapi:deploy:cancel command, refactor base classes to support MDAPI and SOURCE stash keys ([#301](https://github.com/salesforcecli/plugin-source/issues/301)) ([9dcc3e7](https://github.com/salesforcecli/plugin-source/commit/9dcc3e7bc1be13d1dcb0df13b3f7710407f9e977))
+
+### [1.4.3](https://github.com/salesforcecli/plugin-source/compare/v1.4.2...v1.4.3) (2021-11-15)
 
 ### Bug Fixes
 
-* handle a corrupt stash.json file ([#291](https://github.com/salesforcecli/plugin-source/issues/291)) ([ac15a24](https://github.com/salesforcecli/plugin-source/commit/ac15a2433977bbd801fe54e70568bcf176ec0f1b))
+- add back missing help for listmetadata/describemetadata ([#295](https://github.com/salesforcecli/plugin-source/issues/295)) ([f641d83](https://github.com/salesforcecli/plugin-source/commit/f641d8311905d4aa68bcba4679674e4afa8868c1))
+
+### [1.4.2](https://github.com/salesforcecli/plugin-source/compare/v1.4.1...v1.4.2) (2021-11-11)
+
+### Bug Fixes
+
+- handle a corrupt stash.json file ([#291](https://github.com/salesforcecli/plugin-source/issues/291)) ([ac15a24](https://github.com/salesforcecli/plugin-source/commit/ac15a2433977bbd801fe54e70568bcf176ec0f1b))
 
 ### [1.4.1](https://github.com/salesforcecli/plugin-source/compare/v1.4.0...v1.4.1) (2021-11-11)